Add a uiserver ingress template
This adds a uiserver ingress template so that it's possible to create a domain for the uiserver. Support for self-signed
certificates that are directly mounted into the uiserver deployment is being dropped in this commit. Regular deployments
handle certificates through ingresses. Setting TLS in the POD specifically without even having the uiserver exposed is not
as common.

It's true that people could expose the service, a node port, etc. For those cases we will have to come up with a better solution

Signed-off-by: Flavio Percoco <flaper87@gmail.com>
Signed-off-by: yitzchake <yitzchake@reblaze.com>